+++ test/ContentScript/utils.ts
@@ -43,8 +43,9 @@ export function getStaticHttpServer(): http.Server {
 }
 
 export function getFakeZapServer(
-  actualData: Set<string>,
-  JSONPORT: number
+  actualData: Array<string>,
+  JSONPORT: number,
+  incZapEvents = false
 ): http.Server {
   const app = JsonServer.create();
 
@@ -53,9 +54,12 @@ export function getFakeZapServer(
     const action = req.params;
     const {body} = req;
     const msg = JSON.stringify({action, body});
-    actualData.add(
-      msg.replace(/\\"timestamp\\":\d+/g, 'TIMESTAMP').replace(/[\\]/g, '')
-    );
+    if (incZapEvents || msg.indexOf('localzap') === -1) {
+      // Ignore localzap events
+      actualData.push(
+        msg.replace(/\\"timestamp\\":\d+/g, 'TIMESTAMP').replace(/[\\]/g, '')
+      );
+    }
     res.sendStatus(200);
   });