+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 
 ### Changed
-
 - Init ZAP URL to http://zap/ instead of http://localhost:8080/
+
+### Fixed
+- Same links continually reported on domMutation events (Issue 81).
+
 ## 0.0.6 - 2023-09-19
 
 ### Fixed